[turbidity] Hoàn thành việc test - Không dùng uint64_t để tránh lỗi log Serial - Xóa lọc state trong hàm get_stateRaw

+++ TraiCaTam_Code_v2/lib/log/log_service.cpp
@@ -342,7 +342,7 @@ void Log_init()
 {
 #ifdef LOG_PORT
 
-    uint64_t intv = millis();
+    uint32_t intv = millis();
 
     LOG_PORT.begin(LOG_BAUD_RATE);
